Start of login rework
+ New UI (School logo, ...)
+ Async fetching
